Metódo separar
