+++ ProjetoPDF/Renomear.cs
@@ -16,5 +16,25 @@ namespace ProjetoPDF
         {
             InitializeComponent();
         }
+
+        private void lblDestino_Click_1(object sender, EventArgs e)
+        {
+            if (folderBrowser.ShowDialog() != DialogResult.Cancel)
+            {
+                lblDestino.Text = folderBrowser.SelectedPath;
+            }
+        }
+        private void btnRenomear_Click(object sender, EventArgs e)
+        {
+            PDF pdf = new PDF();
+            if (lblDestino.Text != "Clique para selecionar a pasta de destino")
+            {
+                pdf.Renomear(lblDestino.Text);
+            }
+            else
+            {
+                MessageBox.Show("Por favor, selecione o destino");
+            }
+        }
     }
 }
